+++ lib/phoenix_component.ex
@@ -424,16 +424,25 @@ defmodule Phoenix.Component do
   def __global__?(_), do: false
 
   @doc false
+  def __reserved_assigns__, do: [:__changed__, :__slot__, :inner_block, :myself, :flash, :socket]
+
+  @doc false
   defmacro __using__(opts \\ []) do
-    quote bind_quoted: [opts: opts] do
+    helpers =
+      if __CALLER__.module != Phoenix.LiveView.Helpers do
+        quote do: import(Phoenix.LiveView.Helpers)
+      end
+
+    quote do
       import Kernel, except: [def: 2, defp: 2]
       import Phoenix.Component
       import Phoenix.LiveView
-      import Phoenix.LiveView.Helpers
-      global_prefixes = Phoenix.Component.__setup__(__MODULE__, opts)
+      unquote(helpers)
+      global_prefixes = Phoenix.Component.__setup__(__MODULE__, unquote(opts))
       @doc false
       for prefix <- global_prefixes do
-        def __global__?(unquote(prefix) <> _), do: true
+        @phoenix_global_prefix prefix
+        def __global__?(@phoenix_global_prefix <> _), do: true
       end
 
       def __global__?(_), do: false
@@ -703,8 +712,7 @@ defmodule Phoenix.Component do
             nil -> {nil, nil}
           end
 
-        known_keys =
-          for(attr <- attrs, do: attr.name) ++ Phoenix.LiveView.Helpers.__reserved_assigns__()
+        known_keys = for(attr <- attrs, do: attr.name) ++ __reserved_assigns__()
 
         def_body =
           if global_name do