+++ test/phoenix_live_view/integrations/params_test.exs
@@ -117,10 +117,10 @@ defmodule Phoenix.LiveView.ParamsTest do
              |> redirected_to() == "/counter/123?from=rehandled_params"
     end
 
-    test "push_redirect", %{conn: conn} do
+    test "push_navigate", %{conn: conn} do
       assert conn
              |> put_serialized_session(:on_handle_params, fn socket ->
-               {:noreply, LiveView.push_redirect(socket, to: "/thermo/456")}
+               {:noreply, LiveView.push_navigate(socket, to: "/thermo/456")}
              end)
              |> get("/counter/123?from=handle_params")
              |> redirected_to() == "/thermo/456"
@@ -219,12 +219,12 @@ defmodule Phoenix.LiveView.ParamsTest do
       assert response =~ rendered_to_string(~s|mount: %{"from" => "handle_params", "id" => "123"}|)
     end
 
-    test "push_redirect", %{conn: conn} do
+    test "push_navigate", %{conn: conn} do
       {:error, {:live_redirect, %{to: "/thermo/456"}}} =
         conn
         |> put_serialized_session(:on_handle_params, fn socket ->
           if LiveView.connected?(socket) do
-            {:noreply, LiveView.push_redirect(socket, to: "/thermo/456")}
+            {:noreply, LiveView.push_navigate(socket, to: "/thermo/456")}
           else
             {:noreply, socket}
           end
@@ -337,12 +337,12 @@ defmodule Phoenix.LiveView.ParamsTest do
     end
   end
 
-  describe "push_redirect" do
+  describe "push_navigate" do
     test "from event callback", %{conn: conn} do
       {:ok, counter_live, _html} = live(conn, "/counter/123")
 
       assert {:error, {:live_redirect, %{to: "/thermo/123"}}} =
-               render_click(counter_live, :push_redirect, %{to: "/thermo/123"})
+               render_click(counter_live, :push_navigate, %{to: "/thermo/123"})
 
       assert_redirect(counter_live, "/thermo/123")
     end
@@ -353,7 +353,7 @@ defmodule Phoenix.LiveView.ParamsTest do
       next = fn socket ->
         new_socket =
           LiveView.assign(socket, :on_handle_params, fn socket ->
-            {:noreply, LiveView.push_redirect(socket, to: "/thermo/123")}
+            {:noreply, LiveView.push_navigate(socket, to: "/thermo/123")}
           end)
 
         {:reply, :ok, LiveView.push_patch(new_socket, to: "/counter/123?from=handle_params")}
@@ -365,15 +365,15 @@ defmodule Phoenix.LiveView.ParamsTest do
                       %{val: 1}, %{"from" => "handle_params", "id" => "123"}}
     end
 
-    test "shuts down with push_redirect", %{conn: conn} do
+    test "shuts down with push_navigate", %{conn: conn} do
       {:ok, counter_live, _html} = live(conn, "/counter/123")
 
       next = fn socket ->
-        {:noreply, LiveView.push_redirect(socket, to: "/thermo/123")}
+        {:noreply, LiveView.push_navigate(socket, to: "/thermo/123")}
       end
 
       assert {{:shutdown, {:live_redirect, %{to: "/thermo/123"}}}, _} =
-               catch_exit(GenServer.call(counter_live.pid, {:push_redirect, next}))
+               catch_exit(GenServer.call(counter_live.pid, {:push_navigate, next}))
     end
   end