+++ test/support/live_views/components.ex
@@ -71,8 +71,8 @@ defmodule Phoenix.LiveViewTest.StatefulComponent do
       "dup" ->
         {:noreply, assign(socket, :dup_name, socket.assigns.name <> "-dup")}
 
-      "push_redirect" ->
-        {:noreply, push_redirect(socket, to: "/components?redirect=push")}
+      "push_navigate" ->
+        {:noreply, push_navigate(socket, to: "/components?redirect=push")}
 
       "push_patch" ->
         {:noreply, push_patch(socket, to: "/components?redirect=patch")}