+++ test/support/live_views/lifecycle.ex
@@ -21,7 +21,7 @@ defmodule Phoenix.LiveViewTest.MountArgs do
 
   def on_mount(:inlined, _params, _session, socket) do
     qs = URI.encode_query(%{called: true, inlined: true})
-    {:halt, push_redirect(socket, to: "/lifecycle?#{qs}")}
+    {:halt, push_navigate(socket, to: "/lifecycle?#{qs}")}
   end
 end
 
@@ -168,7 +168,7 @@ defmodule Phoenix.LiveViewTest.HooksLive.RedirectMount do
   on_mount __MODULE__
 
   def on_mount(:default, _, _, %{assigns: %{live_action: action}} = socket) do
-    {action, push_redirect(socket, to: "/lifecycle")}
+    {action, push_navigate(socket, to: "/lifecycle")}
   end
 
   def render(assigns), do: ~H"<div></div>"
@@ -190,7 +190,7 @@ defmodule Phoenix.LiveViewTest.HaltConnectedMount do
 
   def on_mount(_arg, _params, _session, socket) do
     if LiveView.connected?(socket) do
-      {:halt, LiveView.push_redirect(socket, to: "/lifecycle")}
+      {:halt, LiveView.push_navigate(socket, to: "/lifecycle")}
     else
       {:cont, LiveView.assign(socket, :last_on_mount, __MODULE__)}
     end